[change] modernize View
Rewrite View to use function components and hooks.

The 'usePlatformMethods' hook also fixes a bug in the class-based implementation of 'setNativeProps' which was unable to correctly merge its styles with those provided via the component API. In the future, 'setNativeProps' will be removed from React Native anyway. See (3) in #1136 for context.